% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histmatch.R
\name{histmatch}
\alias{histmatch}
\alias{histmatch_data}
\title{Histogram matching}
\usage{
histmatch(source, target, w = NULL, open_left = FALSE)

histmatch_data(source, target, w = NULL, open_left = FALSE)
}
\arguments{
\item{source}{\code{[numeric]}\cr Source distribution}

\item{target}{\code{[numeric]}\cr Target distribution}

\item{w}{\code{[numeric]}\cr Optional, a weight vector of the same length as
\code{target}}

\item{open_left}{\code{[logical(1)]}\cr Match against an interval open to the left,
or closed at both ends (default)?}
}
\value{
\itemize{
\item For \code{histmatch()}: A vector of the same length as \code{source}
}

\itemize{
\item For \code{histmatch_data()}: A named list with two components, \code{"source"}
and \code{"target"}, each containing a data frame
}
}
\description{
TBD.
}
\details{
A \code{\link[=plot]{plot()}} method is defined for objects returned by \code{histmatch_data()}.
}

